Доделка задачи
