+++ app/Http/Controllers/Inventory/OrderHistoryController.php
@@ -6,31 +6,33 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Models\Inventory\Order\Order;
 use App\UseCases\Inventory\Order\ExportService;
+use App\UseCases\Inventory\Order\OrderService;
 
 class OrderHistoryController extends Controller
 {
     private $service;
     private $export;
 
-    public function __construct(ExportService $export)
+    public function __construct(OrderService $service, ExportService $export)
     {
         $this->middleware('permission:consumables');
+        $this->service = $service;
         $this->export = $export;
     }
 
     public function index()
     {
-        $orders = Order::paginate(15);
-        return view('inventory.consumables.orderhistory.index', compact('orders'));
+        $orders = $this->service->getPaginate(15);
+        return view('inventory.consumables.orders.history', compact('orders'));
     }
 
     public function destroy(Order $order)
     {
         $order->delete();
-        return redirect()->route('inventory.orderhistory.index');
+        return redirect()->route('inventory.orders.history');
     }
 
-    public function store(Order $order)
+    public function excel(Order $order)
     {
         return $this->export->download($order);
     }
