+++ resources/views/inventory/consumables/orders/history.blade.php
@@ -28,6 +28,9 @@
                                     <td>Ответственное лицо</td>
                                     <td>Пользователь</td>
                                     <td>Объект</td>
+                                    <td>Картридж</td>
+                                    <td>Количество</td>
+                                    <td></td>
                                 </tr>
                             </th>
                         </tr>
@@ -38,18 +41,33 @@
                         <tr>
                             <td><a>{{ $order->responsible }}</a></td>
                             <td><a>{{ $order->user->name }}</a></td>
-                            <td><a>{{ $order->object->code . '-' . $order->object->name }}</a></td>
+                            <td><a>{{ $order->object->getFullName() }}</a></td>
+                            <td>
+                                @foreach ($order->items as $item)
+                                    <ul>
+                                        <li><a>{{ $item->consumable->getFullName() }}</a></li>
+                                    </ul>
+                                @endforeach
+                            </td>
+                            <td>
+                                @foreach ($order->items as $item)
+                                    <ul>
+                                        <li><a>{{ $item->count }}</a></li>
+                                    </ul>
+                                @endforeach
+                            </td>
                             <td class="text-right">
                                 <div class="list-icons-item dropdown">
                                     <a href="#" class="list-icons-item dropdown-toggle caret-0" data-toggle="dropdown" aria-expanded="false"><i class="icon-menu7"></i></a>
                                         <div class="dropdown-menu dropdown-menu-right" x-placement="bottom-end" style="position: absolute; will-change: transform; top: 0px; left: 0px; transform: translate3d(16px, 16px, 0px);">
-                                            <a class="list-icons-item" href="{{ route('inventory.orderhistory.store', $order) }}"><i class="icon-cart2 mr-1"></i>Скачать заказ</a>
+                                            <a class="list-icons-item" href="{{ route('inventory.orderhistory.excel', $order) }}"><i class="icon-file-excel"></i>Скачать заказ</a>
 
                                             <div class="dropdown-divider"></div>
 
                                             <a class="list-icons-item" href="{{ route('inventory.orderhistory.destroy', $order) }}"><i class="icon-cross2"></i>Удалить</a></h6>
                                         </div>
                             </td>
+
                         </tr>
                     @endforeach
                         {{ $orders->links() }}
